requirements.txt
numpy==1.19.2
streamlit==0.71.0
pandas==1.1.3
Pillow==8.0.1
scikit_learn==0.23.2
